+++ tgui/packages/tgui/stories/Blink.stories.js
@@ -1,0 +1,22 @@
+/**
+ * @file
+ * @copyright 2021 Aleksej Komarov
+ * @license MIT
+ */
+
+import { Blink, Section } from '../components';
+
+export const meta = {
+  title: 'Blink',
+  render: () => <Story />,
+};
+
+const Story = (props, context) => {
+  return (
+    <Section>
+      <Blink>
+        Blink
+      </Blink>
+    </Section>
+  );
+};
