+++ tgui/packages/tgui/stories/Stack.stories.js
@@ -1,0 +1,20 @@
+/**
+ * @file
+ * @copyright 2021 Aleksej Komarov
+ * @license MIT
+ */
+
+import { Section } from '../components';
+
+export const meta = {
+  title: 'Stack',
+  render: () => <Story />,
+};
+
+const Story = (props, context) => {
+  return (
+    <Section>
+      asdfasdf
+    </Section>
+  );
+};
